+++ v1_13_R2/src/main/java/net/citizensnpcs/nms/v1_13_R2/entity/EntityHumanNPC.java
@@ -329,7 +329,7 @@ public class EntityHumanNPC extends EntityPlayer implements NPCHolder, Skinnable
         }
     }
 
-    private void moveWithFallDamage(double mx, double my, double mz) {
+    private void moveWithFallDamage(float mx, float my, float mz) {
         double y = this.locY;
 
         a(mx, my, mz);
